refactor: move database module to config folder
